feat: Add docker compose and dockerfiles
