+++ docker/.env.docker
@@ -1,6 +1,11 @@
-# embedding and llm
-OLLAMA_URL=http://localhost:11434
-ES_URL=http://localhost:9200
+# Base Configuration
+PYTHONUNBUFFERED=1
+PYTHONDONTWRITEBYTECODE=1
+PYTHONPATH=/app
+
+# Embedding and LLM
+OLLAMA_URL=http://ollama:11434
+ES_URL=http://elasticsearch:9200
 ES_INDEX=default
 EMBEDDING_MODEL=snowflake-arctic-embed2
 MODEL_NAME=llama3.3
@@ -9,26 +14,34 @@ TEMPERATURE=0.8
 SEED=0
 TOP_K=5
 
-# web
+# Web Configuration
 WEB_DEBUG=false
 WEB_SYSTEM_PROMPT="You are Chipper 🐕, a thoughtful and helpful and professional assistant."
 WEB_CLIENT_DEBUG=true
 
-WEB_HOST=127.0.0.1
+# API Configuration
+WEB_HOST=0.0.0.0
+WEB_API_HOST=api
+WEB_API_PORT=8000
 WEB_CLIENT_PORT=5000
-WEB_API_KEY=DEMO-API-KEY-123
+WEB_API_KEY=MY-API-KEY-123
 API_TIMEOUT=120
 WEB_REQUIRE_SECURE=false
 
-# web asset
+# Asset Configuration
 ASSET_URL=/static
 ASSET_CACHE_TIMEOUT=31536000
 ASSET_DEBUG=False
 
-# web chat
+# Chat Configuration
 MAX_CONTEXT_SIZE=100
 ENABLE_RESPONSE_CACHE=False
 RESPONSE_CACHE_TIMEOUT=300
 
-# logging
+# Elasticsearch Configuration
+ES_JAVA_OPTS=-Xms512m -Xmx512m
+DISCOVERY_TYPE=single-node
+XPACK_SECURITY_ENABLED=false
+
+# Logging
 LOG_LEVEL=INFO
