+++ docker/dockerfiles/Dockerfile.webclient
@@ -1,0 +1,33 @@
+FROM python-base AS builder
+
+WORKDIR /app
+
+COPY src/web/web_client_server.py ./src/web/
+COPY src/web/static ./src/web/static
+COPY src/web/templates ./src/web/templates
+
+ARG BUILD_ENV=production
+RUN if [ "$BUILD_ENV" = "production" ]; then \
+    pip install --no-cache-dir gunicorn; \
+    fi
+
+FROM python:3.12-slim
+
+WORKDIR /app
+
+COPY --from=builder /opt/venv /opt/venv
+ENV PATH="/opt/venv/bin:$PATH"
+
+COPY --from=builder /app/src ./src
+ENV PYTHONPATH=/app
+
+RUN useradd -m -u 1000 webuser && \
+    chown -R webuser:webuser /app
+USER webuser
+
+EXPOSE 5000
+
+HEALTHCHECK --interval=30s --timeout=3s --start-period=30s --retries=3 \
+    CMD curl -f http://localhost:5000/api/local-health || exit 1
+
+CMD ["gunicorn", "--bind", "0.0.0.0:5000", "--workers", "4", "--threads", "2", "--timeout", "120", "src.web.web_client_server:app"]
