+++ src/web/rest_server.py
@@ -26,10 +26,10 @@ limiter = Limiter(
     key_func=get_remote_address,
     app=app,
     default_limits=["1000 per day", "10 per minute"],
-    storage_uri="memory://"
+    storage_uri="memory://",
 )
 
-API_KEY = os.getenv('WEB_API_KEY')
+API_KEY = os.getenv("WEB_API_KEY")
 if not API_KEY:
     API_KEY = secrets.token_urlsafe(32)
     logger.info(f"Generated API key: {API_KEY}")
@@ -38,10 +38,10 @@ if not API_KEY:
 def process_query():
     try:
         data = request.get_json()
-        query = data.get('query')
-        es_index = data.get('es_index')
-        model_name = data.get('model_name')
-        conversation = data.get('conversation', [])
+        query = data.get("query")
+        es_index = data.get("es_index")
+        model_name = data.get("model_name")
+        conversation = data.get("conversation", [])
 
         rag = RAGPipeline(
             es_index=es_index,
@@ -52,9 +52,7 @@ def process_query():
         result = None
         try:
             result = rag.run_query(
-                query=query,
-                conversation=conversation,
-                print_response=False
+                query=query, conversation=conversation, print_response=False
             )
         except Exception as e:
             success = False
@@ -62,31 +60,44 @@ def process_query():
         latest_interaction = []
         if success:
             latest_interaction = [
-                {"role": "user", "content": query,
-                 "timestamp": datetime.now().isoformat()},
-                {"role": "assistant", "content": result["llm"]["replies"][0],
-                 "timestamp": datetime.now().isoformat()}
+                {
+                    "role": "user",
+                    "content": query,
+                    "timestamp": datetime.now().isoformat(),
+                },
+                {
+                    "role": "assistant",
+                    "content": result["llm"]["replies"][0],
+                    "timestamp": datetime.now().isoformat(),
+                },
             ]
 
-        return jsonify({
-            "success": success,
-            "timestamp": datetime.now().isoformat(),
-            "result": result,
-            "conversation": conversation + latest_interaction
-        })
+        return jsonify(
+            {
+                "success": success,
+                "timestamp": datetime.now().isoformat(),
+                "result": result,
+                "conversation": conversation + latest_interaction,
+            }
+        )
 
     except Exception as e:
         logger.error(f"Error processing query: {str(e)}", exc_info=True)
-        return jsonify({
-            "success": False,
-            "error": "An error occurred while processing your request"
-        }), 500
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": "An error occurred while processing your request",
+                }
+            ),
+            500,
+        )
 
 
 def require_api_key(f):
     @wraps(f)
     def decorated_function(*args, **kwargs):
-        api_key = request.headers.get('X-API-Key')
+        api_key = request.headers.get("X-API-Key")
         if not api_key or api_key != API_KEY:
             return jsonify({"error": "Invalid or missing API key"}), 401
         return f(*args, **kwargs)
@@ -97,42 +108,44 @@ def require_api_key(f):
 @app.before_request
 def before_request():
     logger.info(f"Request {request.method} {request.path} from {request.remote_addr}")
-    if os.getenv('WEB_REQUIRE_SECURE', 'False').lower() == 'true' and not request.is_secure:
+    if (
+        os.getenv("WEB_REQUIRE_SECURE", "False").lower() == "true"
+        and not request.is_secure
+    ):
         return jsonify({"error": "HTTPS required"}), 403
 
 
 @app.after_request
 def after_request(response):
-    response.headers['Strict-Transport-Security'] = 'max-age=31536000; includeSubDomains'
-    response.headers['X-Content-Type-Options'] = 'nosniff'
-    response.headers['X-Frame-Options'] = 'DENY'
-    response.headers['X-XSS-Protection'] = '1; mode=block'
+    response.headers["Strict-Transport-Security"] = (
+        "max-age=31536000; includeSubDomains"
+    )
+    response.headers["X-Content-Type-Options"] = "nosniff"
+    response.headers["X-Frame-Options"] = "DENY"
+    response.headers["X-XSS-Protection"] = "1; mode=block"
     return response
 
 
-@app.route('/api/query', methods=['POST'])
+@app.route("/api/query", methods=["POST"])
 @require_api_key
 def query_endpoint():
     return process_query()
 
 
-@app.route('/api/embed', methods=['POST'])
+@app.route("/api/embed", methods=["POST"])
 @require_api_key
 def embed_endpoint():
     return process_embedding()
 
 
-@app.route('/api/health', methods=['GET'])
+@app.route("/api/health", methods=["GET"])
 def health_check():
-    return jsonify({
-        "status": "healthy",
-        "timestamp": datetime.utcnow().isoformat()
-    })
+    return jsonify({"status": "healthy", "timestamp": datetime.utcnow().isoformat()})
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     app.run(
-        host=os.getenv('WEB_HOST', '0.0.0.0'),
-        port=int(os.getenv('WEB_PORT', '5001')),
-        debug=os.getenv('WEB_DEBUG', 'False').lower() == 'true'
+        host=os.getenv("WEB_API_HOST", "0.0.0.0"),
+        port=int(os.getenv("WEB_API_PORT", "8000")),
+        debug=os.getenv("WEB_DEBUG", "False").lower() == "true",
     )
